Make multiplication output u8 matrix

+++ src/quantization.rs
@@ -1,6 +1,6 @@
 pub trait Quantizer4Bit {
     fn quantize(&self, real_val: f32) -> u8;
-    fn dequantize(&self, q_val: i32) -> f32;
+    fn dequantize(&self, q_val: u8) -> f32;
 }
 
 pub fn quantize_and_pack(quantizer: &impl Quantizer4Bit, v1: f32, v2: f32) -> u8 {
@@ -67,7 +67,7 @@ impl Quantizer4Bit for AffineQuantizer {
             .round() as u8
     }
 
-    fn dequantize(&self, q_val: i32) -> f32 {
+    fn dequantize(&self, q_val: u8) -> f32 {
         self.scale * (q_val as f32 - self.zero as f32)
     }
 }
